Lab3_FastAPI/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
xgboost
seaborn
google-cloud-storage
python-multipart